PLT-7583 Consolided validator experiments into proposed new validator for production.
