+++ changelog.d/20231209_094720_brian.bush_PLT_7583.md
@@ -1,0 +1,4 @@
+### Changed
+
+- The `Contract` and `Action` types now use `PlutusTx.asData`.
+- The `PlutusTx.geq` comparison function has been specialized.
