+++ marlowe-plutus/app/Benchmark/Marlowe.hs
@@ -24,7 +24,7 @@ import Control.Monad.Writer (runWriterT)
 import Data.Bifunctor (bimap)
 import Data.Either.Extras (unsafeFromEither)
 import Data.List (isSuffixOf)
-import Language.Marlowe.Core.V1.Semantics (MarloweData)
+import Language.Marlowe.Plutus.Semantics (MarloweData)
 import Language.Marlowe.Scripts.Types (MarloweInput)
 import Paths_marlowe_plutus (getDataDir)
 import PlutusCore.Executable.AstIO (fromNamedDeBruijnUPLC)
