--- marlowe-plutus/src/Language/Marlowe/Plutus/AsData/Semantics/Types/Address.hs
@@ -1,19 +1,0 @@
------------------------------------------------------------------------------
---
--- Module      :  $Headers
--- License     :  Apache 2.0
---
--- Stability   :  Experimental
--- Portability :  Portable
---
-
------------------------------------------------------------------------------
-
--- | Address types for Marlowe.
-module Language.Marlowe.Plutus.AsData.Semantics.Types.Address (
-  -- * Types
-  Network,
-) where
-
--- | Type of network.
-type Network = Bool
